requirements.txt
antspyx==0.6.1
captum==0.8.0
joblib==1.4.2
lightning==2.5.1
lxt==2.1
monai==1.4.0
nibabel==5.3.2
nilearn==0.11.1
numpy==1.26.4
pandas==2.2.3
pillow==11.1.0
python-dotenv==1.2.1
PyYAML==6.0.2
SimpleITK==2.4.1
SimpleITK==2.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
timm==1.0.15
timm-3d==1.0.1
torch==2.6.0
torchmetrics==1.7.1
torchvision==0.21.0
tqdm==4.67.1
wandb==0.14.0
zennit==0.5.1
